Add activity for viewing all sources that contributed to a game
This change includes what looks like a duplicate set of activities:
GameListActivity and GameDetailActivity.

Those are going to be the future home of the current activities, and
they support a nice fragment UI approach that will look good on
tablets. But for now I just want to get some working code submitted so
I can carefully diff changes when I make that transition.